Provide a set of button labels with glyphs
